fix bug (saving an edited recipe was not possible)

+++ recapi/models/tagmodel.py
@@ -30,31 +30,34 @@ class RecipeTags(BaseModel):
 
 def add_tags(recipe_data, recipe_id):
     """Add entries for Tag and RecipeTags and delete removed tags."""
-    newTags = recipe_data.get("newTags", {})
-    for tagname, category in newTags.items():
-        tagname = tagname.lower().strip()
-        tag = Tag(tagname=tagname, parent=TagCategory.get(TagCategory.categoryname == category))
-        tag.save()
+    newTags = recipe_data.get("newTags") or {}
+    if isinstance(newTags, dict):
+        for tagname, category in newTags.items():
+            tagname = tagname.lower().strip()
+            tag = Tag(tagname=tagname, parent=TagCategory.get(TagCategory.categoryname == category))
+            tag.save()
 
     # Get existing tags for recipe
     existing_tags_rows = RecipeTags.select().join(Tag, pw.JOIN.LEFT_OUTER).where((RecipeTags.recipeID == recipe_id))
     existing_tags = []
     for row in existing_tags_rows:
         row_dict = model_to_dict(row)
-        if row_dict.get("tagID", {}).get("tagname"):
-            existing_tags.append(row_dict.get("tagID", {}).get("tagname"))
+        tag_dict = row_dict.get("tagID")
+        if isinstance(tag_dict, dict) and tag_dict.get("tagname"):
+            existing_tags.append(tag_dict["tagname"])
 
     # Add tags if they don't exist already
-    tags = recipe_data.get("tags", [])
-    for tagname in tags:
-        tagname = tagname.lower().strip()
-        if tagname not in existing_tags:
-            recipetags = RecipeTags(recipeID=Recipe.get(Recipe.id == recipe_id), tagID=Tag.get(Tag.tagname == tagname))
-            recipetags.save()
+    tags = recipe_data.get("tags") or []
+    if isinstance(tags, list):
+        for tagname in tags:
+            tagname = tagname.lower().strip()
+            if tagname not in existing_tags:
+                recipetags = RecipeTags(recipeID=Recipe.get(Recipe.id == recipe_id), tagID=Tag.get(Tag.tagname == tagname))
+                recipetags.save()
 
     # Delete removed tags
     for tagname in existing_tags:
-        if tagname not in tags:
+        if isinstance(tags, list) and tagname not in tags:
             recipetags = RecipeTags.get(RecipeTags.recipeID == recipe_id, RecipeTags.tagID == Tag.get(Tag.tagname == tagname).id)
             recipetags.delete_instance()
             # Remove this tag from Tag table if no other recipe uses it
